+++ archiverse/__init__.py
@@ -1,0 +1,4 @@
+from .archiverse import main
+
+__all__ = ["main"]
+